+++ go.work
@@ -1,4 +1,4 @@
-go 1.22.2
+go 1.20
 
 use (
 	./hw02_fix_app
@@ -12,8 +12,7 @@ use (
 	./hw10_motion_sensor
 	./hw11_worker_pool
 	./hw12_log_util
-	./hw13_http/server
-	./hw13_http/client
+	./hw13_http
 	./hw15_go_sql
 	./hw16_docker
 )
