+++ hw13_http/client/client_test.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"testing"
