--- hw13_http/client/go.mod
@@ -1,5 +1,0 @@
-module github.com/kvs4/home_work_basic/hw13_http/client
-
-go 1.22.2
-
-require github.com/spf13/pflag v1.0.5
